restruct plots etc to be under run/ dir so we have past results

+++ run_make_network.sh
@@ -4,11 +4,16 @@ if [[ -z "$WAVE" ]]; then
   exit 1
 fi
 
+if [[ -z "$RUN" ]]; then
+  echo "RUN not set"
+  exit 1
+fi
+
 set -x
 rm -rf sverilog_version/tests/network/{net.out,test_x.hex} y_pred.sverilog.txt verilog.y_pred.png
 
 set -e
-cp test_x*hex sverilog_version/tests/network/
+cp runs/$RUN/test_x_files/test_x*hex sverilog_version/tests/network/
 
 # run iverilog sim
 pushd sverilog_version/tests/network
@@ -17,8 +22,9 @@ make | tee net.$WAVE.out
 popd
 
 # generate plot
-cat sverilog_version/tests/network/net.$WAVE.out \
+mv sverilog_version/tests/network/net.$WAVE.out runs/$RUN/
+cat runs/$RUN/net.$WAVE.out \
  | grep "^OUT dec" | cut -f3 -d' ' | grep -v xxxx | uniq \
- > y_pred.sverilog.$WAVE.txt
-./plot.py --plot-png verilog.y_pred.$WAVE.png < y_pred.sverilog.$WAVE.txt
-rm y_pred.sverilog.$WAVE.txt
+ > /tmp/$$.y_pred.sverilog.$WAVE.txt
+./plot.py --plot-png runs/$RUN/verilog.y_pred.$WAVE.png < /tmp/$$.y_pred.sverilog.$WAVE.txt
+rm /tmp/$$.y_pred.sverilog.$WAVE.txt
